requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.1.3
django-browser-reload==1.17.0
django-crispy-forms==2.3
idna==3.10
pillow==11.0.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0

# nvm install lts
# nvm use lts
